V0.4.3
NAPS System
Administration Users
- Created the methods of add_user, update_user and delete_user on the
admin_user controller to manage the users of the database
- Created the same methods in the model but with the necessary queries
to represent the data updating
Notes:
*Add User validates if the mail already exists because that's the login
id
*Update User validates if your mail is different that the previous you
had and if it is... that the person editing doesn't put a existing
mail.
Ex: Registry 1 = davidcastro@outlook.com, Registry 2 =
iespinoza@tekmexico.com
I choose to edit Registry 2 and assign a new mail which is going to be
"davidcastro@outlook.com" since a registry with the previous mail
already exists, it will fail.
- Modified the query of the Sign_in to only select the active users
even if the credentials of the user are right. That means that if an
administrator is inactive, he/she won't be able to log in.
- Included some HTML validation and new patterns on the name, last name
and password field, and some titles too.
- Created the handlers for the add/edit/delete and also the save.
Several validations against typing wrong passwords/ leaving password
blank/ etc etc.
TODO:
- Comment Everything.
- Automatic Refresh: Right now i'm refreshing the page completely on
every action like delete/update/add; Next thing to do is to refresh the
table automatically without refreshing the whole page.