V0.5.2
NAPS
- Users not logged are able to see the presentations but not evaluate
- Users logged can evaluate and the button is marked on blue; Whenever
the user presses that button a flip form will appear to let the user
evaluate the presentation with stars and comments (Still missing
validation to not let user evaluate with 0 stars)
- Whenever the user evaluated someone, the page refreshes automatically
putting the evaluated users in green and not letting the user
re-evaluate a user. The good thing about this is that since we're using
PHP Server Side Technology at the moment of create the user to display
we can choose  which set of HTML show and that means that if an user
was evaluated when we construct the HTML it will never construct the
form too and hide it, so there is no way to make the form appear and
try to vote again :)
- Added a Link to the user name; This will be the part where you can
click the user name and navigate to that user's page where their score
and comments will be displayed :P
So everyone can spy each other!
- Changed the database structure in Workbench